+++ src/organizadorEco/PendientePanel.java
@@ -1,0 +1,148 @@
+package organizadorEco;
+
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.*;
+
+public class PendientePanel extends JPanel implements MouseListener, ItemListener {
+    final int WIDTH = 300;
+    final int HEIGHT = 40;
+    final int SIZE = 15;
+    int day, month, year;
+    JLabel label;
+    JTextArea area;
+    JButton eliminar;
+    JButton confirmar;
+    JButton guardar;
+    JComboBox<Integer> days;
+    JComboBox<String> months;
+    JComboBox<Integer> years;
+
+    public PendientePanel(String descripcion, int a, int m, int d) {
+        this.setLayout(new FlowLayout(FlowLayout.CENTER, 30, 10));
+        this.setPreferredSize(new Dimension(WIDTH, HEIGHT));
+        this.setBackground(GUI.colorTerciario);
+
+        days = new JComboBox<>();
+        months = new JComboBox<>();
+        years = new JComboBox<>();
+        for (int i = 1; i <= 31; i++) {
+            if (i <= 12) this.months.addItem(Calendario.mes[i - 1]);
+            if (i <= 10) this.years.addItem((2019 + i));
+            this.days.addItem(i);
+        }
+        this.days.addItemListener(this);
+        this.months.addItemListener(this);
+        this.years.addItemListener(this);
+
+        day = d;
+        month = m;
+        year = a;
+
+        days.setSelectedItem(day);
+        months.setSelectedItem(Calendario.mes[month - 1]);
+        years.setSelectedItem(year);
+
+        label = new JLabel(descripcion, SwingConstants.CENTER);
+        label.setFont(new Font(GUI.fuente, Font.PLAIN, SIZE));
+        label.setVerticalAlignment(SwingConstants.CENTER);
+        label.setPreferredSize(new Dimension(WIDTH - 15, HEIGHT / 2));
+
+        area = new JTextArea();
+        area.setFont(new Font(GUI.fuente, Font.PLAIN, SIZE));
+        area.setText(descripcion);
+        area.setWrapStyleWord(true);
+        area.setLineWrap(true);
+        area.setPreferredSize(new Dimension(280, 95));
+
+        eliminar = new JButton();
+        ImageIcon remove = new ImageIcon("imagenes/remove.png");
+        eliminar.setIcon(remove);
+        eliminar.setBackground(null);
+        eliminar.setBorder(null);
+        eliminar.addActionListener(e -> {
+            Organizador.modificarPendiente(descripcion, area.getText(), year, month, day);
+            GUI.deleted.borrar(area.getText());
+            GUI.task.actualizarPaneles();
+            GUI.calendar.actualizarPaneles();
+        });
+
+        confirmar = new JButton();
+        ImageIcon checkmark = new ImageIcon("imagenes/checkmark.png");
+        confirmar.setIcon(checkmark);
+        confirmar.setBackground(null);
+        confirmar.setBorder(null);
+        confirmar.addActionListener(e -> {
+            Organizador.modificarPendiente(descripcion, area.getText(), year, month, day);
+            GUI.done.agregar(area.getText());
+            GUI.task.actualizarPaneles();
+            GUI.calendar.actualizarPaneles();
+        });
+
+        guardar = new JButton();
+        ImageIcon save = new ImageIcon("imagenes/save.png");
+        guardar.setIcon(save);
+        guardar.setBackground(null);
+        guardar.setBorder(null);
+        guardar.addActionListener(e -> {
+            this.remove(area);
+            this.remove(days);
+            this.remove(months);
+            this.remove(years);
+            this.remove(confirmar);
+            this.remove(eliminar);
+            this.remove(guardar);
+            this.add(label);
+            this.setPreferredSize(new Dimension(WIDTH, HEIGHT));
+            label.setText(area.getText());
+            Organizador.modificarPendiente(descripcion, area.getText(), year, month, day);
+            GUI.task.actualizarPaneles();
+            GUI.calendar.actualizarPaneles();
+            revalidate();
+            repaint();
+        });
+
+        this.add(label);
+        this.addMouseListener(this);
+    }
+
+    @Override
+    public void mouseClicked(MouseEvent e) {
+        this.setPreferredSize(new Dimension(WIDTH, HEIGHT + 155));
+        this.remove(label);
+        this.add(area);
+        this.add(days);
+        this.add(months);
+        this.add(years);
+        this.add(eliminar);
+        this.add(confirmar);
+        this.add(guardar);
+        revalidate();
+        repaint();
+    }
+
+    @Override
+    public void mousePressed(MouseEvent e) { }
+
+    @Override
+    public void mouseReleased(MouseEvent e) { }
+
+    @Override
+    public void mouseEntered(MouseEvent e) { }
+
+    @Override
+    public void mouseExited(MouseEvent e) { }
+
+    @Override
+    public void itemStateChanged(ItemEvent e) {
+        if (e.getSource() == this.days) {
+            this.day = (Integer)this.days.getSelectedItem();
+        }
+        if (e.getSource() == this.months) {
+            this.month = this.months.getSelectedIndex() + 1;
+        }
+        if (e.getSource() == this.years) {
+            this.year = (Integer)this.years.getSelectedItem();
+        }
+    }
+}
